refactor(auth): refactored file structure
