+++ src/auth/oauth2/models/index.ts
@@ -1,0 +1,3 @@
+export { default as oauth2 } from './oauth2.class'
+export { OAuth2Metadata } from './OAuth2Metadata.interface'
+export { XUIOAuth2Strategy } from './XUIOAuth2Strategy.class'
