+++ src/auth/oidc/models/index.ts
@@ -1,0 +1,2 @@
+export { default as oidc } from './openid.class'
+export { OpenIDMetadata } from './OpenIDMetadata.interface'
