+++ src/auth/oidc/models/openid.class.spec.ts
@@ -1,4 +1,4 @@
-import oidc from './index'
+import oidc from './openid.class'
 
 test('OIDC Auth', () => {
     expect(oidc).toBeDefined()
