+++ src/auth/oidc/validation/index.ts
@@ -1,0 +1,1 @@
+export { ValidateOpenIdOptions } from './openIdOptions.validation'
